+++ src/main/kotlin/no/nav/hjelpemidler/delbestilling/Application.kt
@@ -10,7 +10,7 @@ import io.ktor.server.routing.routing
 import no.nav.hjelpemidler.delbestilling.delbestilling.azureRoutes
 import no.nav.hjelpemidler.delbestilling.delbestilling.delbestillingApiAuthenticated
 import no.nav.hjelpemidler.delbestilling.delbestilling.delbestillingApiPublic
-import no.nav.hjelpemidler.delbestilling.delbestilling.medDelbestillerRolle
+import no.nav.hjelpemidler.delbestilling.plugins.medDelbestillerRolle
 import no.nav.hjelpemidler.hjelpemidler.hjelpemidler.hjelpemiddelApi
 import no.nav.tms.token.support.azure.validation.AzureAuthenticator
 import no.nav.tms.token.support.tokenx.validation.TokenXAuthenticator
@@ -30,9 +30,9 @@ fun Application.setupRoutes() {
     routing {
         route("/api") {
             authenticate(TokenXAuthenticator.name) {
-                medDelbestillerRolle {
-                    delbestillingApiAuthenticated(ctx.delbestillingService)
-                }
+                medDelbestillerRolle(ctx.rolleService)
+
+                delbestillingApiAuthenticated(ctx.delbestillingService)
             }
 
             rateLimit(RateLimitName("public")) {
