+++ src/main/kotlin/no/nav/hjelpemidler/delbestilling/delbestilling/DelbestillingApi.kt
@@ -2,7 +2,6 @@ package no.nav.hjelpemidler.delbestilling.delbestilling
 
 import io.ktor.http.HttpStatusCode
 import io.ktor.server.application.call
-import io.ktor.server.application.install
 import io.ktor.server.request.receive
 import io.ktor.server.response.respond
 import io.ktor.server.routing.Route
@@ -11,7 +10,6 @@ import io.ktor.server.routing.post
 import io.ktor.server.routing.put
 import io.ktor.server.util.getOrFail
 import mu.KotlinLogging
-import no.nav.hjelpemidler.delbestilling.plugins.DelbestillerRollePlugin
 import no.nav.hjelpemidler.delbestilling.tokenXUser
 
 private val log = KotlinLogging.logger {}
@@ -85,9 +83,4 @@ fun Route.azureRoutes(
         call.respond(HttpStatusCode.OK)
         log.info { "Status for delbestilling $id (hmdel_$id) oppdatert OK" }
     }
-}
-
-fun Route.medDelbestillerRolle(routes: Route.() -> Unit) = createChild(selector).apply {
-    install(DelbestillerRollePlugin)
-    routes()
 }
